requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
fastapi==0.115.12
fonttools==4.57.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
kombu==5.5.3
limits==5.1.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==21.0.1
pytz==2025.2
redis==5.3.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
tqdm
